-- ATTEMPTED to create a StatusEffect class and sub-classes in order to get rid of the  Cold / Fire code in the Enemy.cs file (seemed kind of out of place and didn't lend itself to re-usability, plus after all status effects are created this would add huge bulk to Enemy class). To facilitate this change, I wanted to add a list of StatusEffects that were added to the Enemy each time he was hit, and in the Enemy's update function it would loop through all them and take the appropriate actions.
However, it seems moving the  StartCoRoutine("FireTick", 1.0f) is causing issues with Null References. What's null in this case? It's not the Enemy being affected by the status; it's something to do with the CoRoutine, as when I comment it out everything works without bugs (though no status effects are applied obviously).

Would be great to get some help with this because I'm not very knowledgeable with Coroutines.